requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9        # PostgreSQL driver (useful if you switch from SQLite)
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4
alembic==1.13.2               # For DB migrations
